Create AbstractLogContext class with static and dynamic subclasses, remove DynamicPrefixLogger
